Improves handling of dialogs for upload, delete, download and create/edit AAS
