+++ aas-web-ui/src/components/EditorComponents/AAS.vue
@@ -1,25 +1,8 @@
 <template>
-    <v-tooltip open-delay="600" location="end">
-        <template #activator="{ props }">
-            <v-list-item v-if="newAAS" slim v-bind="props" @click="editAASDialog = true">
-                <template #prepend>
-                    <v-icon size="small">mdi-plus</v-icon>
-                </template>
-                Create AAS
-            </v-list-item>
-            <v-list-item v-else @click="editAASDialog = true">
-                <template #prepend>
-                    <v-icon size="x-small">mdi-pencil</v-icon>
-                </template>
-                <v-list-item-subtitle>Edit AAS</v-list-item-subtitle>
-            </v-list-item>
-        </template>
-        <span>Creat a new AAS</span>
-    </v-tooltip>
     <v-dialog v-model="editAASDialog" width="860" persistent>
         <v-card>
             <v-card-title>
-                <span class="text-subtile-1">{{ newAAS ? 'Create a new AAS' : 'Edit AAS' }}</span>
+                <span class="text-subtile-1">{{ newShell ? 'Create a new AAS' : 'Edit AAS' }}</span>
             </v-card-title>
             <v-divider></v-divider>
             <v-card-text style="overflow-y: auto">
@@ -81,12 +64,17 @@
 
 <script lang="ts" setup>
     import { types as aasTypes } from '@aas-core-works/aas-core3.0-typescript';
-    import { ref } from 'vue';
+    import { ref, watch } from 'vue';
     import { useAASRepositoryClient } from '@/composables/Client/AASRepositoryClient';
     import { UUID } from '@/utils/IDUtils';
 
     const props = defineProps<{
-        newAAS: boolean;
+        modelValue: boolean;
+        newShell: boolean;
+    }>();
+
+    const emit = defineEmits<{
+        (event: 'update:modelValue', value: boolean): void;
     }>();
 
     const { postAas } = useAASRepositoryClient();
@@ -110,9 +98,23 @@
     const globalAssetId = ref<string | null>(null);
     const assetType = ref<string | null>(null);
 
+    watch(
+        () => props.modelValue,
+        (value) => {
+            editAASDialog.value = value;
+        }
+    );
+
+    watch(
+        () => editAASDialog.value,
+        (value) => {
+            emit('update:modelValue', value);
+        }
+    );
+
     function saveAAS() {
         if (AASId.value === null) return;
-        if (props.newAAS) {
+        if (props.newShell) {
             // Create new Asset Information object
             const assetInformation = new aasTypes.AssetInformation(assetKind.value);
